An alternative EPUB generator.
This script generates the office texts in EPUB format for one or several years.
Resulting files can easily be converted to MOBI format using kindlegen.

It does not rely on ebook-convert to do the conversions, but the XHTML files
(necessary for the EPUB format) are generated using customized perl scripts.
Consequently, the resulting files are considerably smaller and the production
is faster.

Up-to-date generated files using 1962 rubrics are maintained on:
http://breviarium.srubarovi.cz

Relevant for issues #36 and #206.