Partial fix of  #411 (Hungarian)
I've added the correct verse for September 14.
Source of translation: http://breviar.sk/include_hu/sv_sep.htm

I haven't managed to find a verse for May 3, so I've made it to display as it used to be; however, I think that the version from September 14 would be more fitting (still incorrect, though).