Update Pasc2-0.txt
The eighth reading (in Reduced 55, Divino Afflatu, etc) for the Second Sunday after Pentecost concluded with this text: "Antiphon at the Song of Zacharias. I am the Shepherd of the sheep: * I am the Way, the Truth, and the Life: I am the Good Shepherd, and know My sheep, and am known of Mine. Alleluia, Alleluia."
This was clearly an inadvertent copy/paste or similar error, so I deleted it.