1960 Newcalendar: Prefer feasts specified in 'Cum Sanctissima' to Lenten ferias.
The logic to force the rank of Lenten/Passiontide ferias to 4.9 was redundant,
predating the ability to set ranks conditionally according to rubrics in the
datafiles.  Having removed this, the datafiles of the relevant feasts have been
changed so that they maintain their pre-1960 rank that allows them to beat such
ferias.