Fixed some spelling + grammar
Previous: Pope Leo XIII on 20 December 1884, allowed one year indulgence for any priest saying the Canticle Tres pueri, Psalm 150 and the accompanied antiphons, vesicles and prayers and the prayers of St Thomas Aquinas and St. Bonaventure as below.

Now: Pope Leo XIII on the 20th of December 1884, allowed a one year indulgence for any priest saying the Canticle Tres pueri, Psalm 150 and the accompanied antiphons, versicles and prayers, and the prayers of Saint Thomas Aquinas and Saint Bonaventure as below.