Addressing the deeper cause of issue #3095
S. Monica had no Ant 2. But not because of the cleanup of C7a but because of the double cross-references I had introduced in C7ap.

Double checked with S. Monica (May 4, 2022 for Rubrics1960 and 2021 for DivinoAfflatu) that it works now in T. P.; and with S. Birgitta (October 8) also outside of T.P.