Dynamic Dominica infra 8vam Epiphania
If January 6 and hence January 13 is a Sunday, then the office of the Sunday (including Holy Family) is anticipated to Saturday January 12.
In addition, according to the Rubrics 1888, there are only 6 days of the Octave as the provided Scripture of the letter to the Romans always terminates before the Sunday. Hence, the days in the Octave need to be transfered by one day each for pre Divino rubrics.
Transfer tables for dominical letters other than f are going to follow in the next commit.