Octaves and Ranking of Octaves and Ranking of Days and some Code Comments
1. Octaves and Ranking of days
The solutions to Issues #3269 #3284 and #3288 concerning the ranking and labelling of Octaves is incomplete as, e.g., the CC Octave fails to outrank S. Barnabas Ap. in 2012 despite being an Octave of II. Order.

This commit addresses these issues comprehensively.
In addition it solves new issue #3308

2. Dynamic Dominica infra 8vam Epiphania
If January 6 and hence January 13 is a Sunday, then the office of the Sunday (including Holy Family) is anticipated to Saturday January 12.
In addition, according to the Rubrics 1888, there are only 6 days of the Octave as the provided Scripture of the letter to the Romans always terminates before the Sunday. Hence, the days in the Octave need to be transfered by one day each for pre-Divino rubrics depending on the dominical letter.
Transfer tables are going to follow in the next commit.