Ranking of Anticipated Sundays and their 1st Nocturn lessons
Under Tridentine Rubrics, anticipated sundays are simplex and are not be placed on a day of nine lessons, if possible. the only nocturn contains the homily of nocturn 3. For the case that Septuag = Epi2 there is a special rubric which can only be followed by the word under 1570 rubrics. Under 1910 there remains a conflict as S. Hiiarius has been raised to Duplex but the rubric presumes it's a semiduplex.

Under divino rubrics, the anticipated sunday is always placed on the preceding saturday under the rite of semiduplex but with all the previliges of minor sundays. in addition, if there is an initia on that sunday its anticipated to the friday and the initia of the sunday is to be accompanied by the additional initias set out in the anticipated week, if any.