Officium of the Dead; All Souls when on a Sunday
When November 2 is a Sunday, i.e. Dominical Letter e, the Officium of the dead is to be transferred to the following Monday. Since for Nov 2 Vespers, the winner is still the Temporal office, the mechanism in place when All Souls is commemorated on the Feria, i.e., via the [Rule] in the Sancti 11-01 and 11-02 files don't work.
The easier way to make the algorithm put the Office of the Dead on the right days is by using the transfer files as done here and before and unfortunately deleted in the recent dirge update commit #3313 .
The alternative would need to create 5 different additional Sunday files in the Tempora and use them in the transfer files anyways.