Update CompletoriumOP.txt
Further changes to bring the style and formatting of the English CompletoriumOP into line with the rest of the formatting for the other Completorium 'editions' (Roman, etc). 

Changes were also made to try and use English translations not under copyright, as well.  I think the 1967 English translation of the Dominican Office is still under copyright, so I was trying to avoid any issues on that front.