Update CompletoriumOP.txt
This commit cleans up the CompletoriumOP file and matches the Latin to the 1962 Dominican (Latin) Breviary.  

The Marian antiphon prayer is wrong (Omnipotens is not the prayer, but rather Concede nos).  The versicle/response for the Salve is also wrong (Ora pro nobis vs. Dignare me laudare te).

Also brings formatting (notably the first characters, etc) in line with the rest of the site.  Also includes the termination of Completorium (Fidelium animae) which was omitted.