docker-compose.yml: Set private SELinux re-labeling
On systems with SELinux enabled, it is required to set correct labels
on the mounted directories. Otherwise, the container does not have
access to the directory. This results in the "(13)Permission denied"
errors from the Apache server.

Set the SELinux private re-labeling option on the mounted directory
to fix the issue.

The SELinux re-labeling option is ignored on platforms without SELinux,
so it should not break already existing setups.

Tested on Fedora 41 with podman 5.4.2.

[1] https://docs.docker.com/reference/compose-file/services/#short-syntax-5